# temp-swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "NewsAI",
    "version": "1.0"
  },
  "paths": {
    "/api/Configuracion/{usuarioId}": {
      "get": {
        "tags": [
          "Configuracion"
        ],
        "parameters": [
          {
            "name": "usuarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Configuracion"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Configuracion"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Configuracion"
                }
              }
            }
          }
        }
      }
    },
    "/api/Configuracion": {
      "post": {
        "tags": [
          "Configuracion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Configuracion"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Configuracion"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Configuracion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Configuracion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "hashtags": {
            "type": "string",
            "nullable": true
          },
          "gradoDesarrolloResumen": {
            "type": "string",
            "nullable": true
          },
          "lenguaje": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "boolean"
          },
          "audio": {
            "type": "boolean"
          },
          "usuarioId": {
            "type": "integer",
            "format": "int32"
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          }
        },
        "additionalProperties": false
      },
...
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "login": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "fechaAlta": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "configuraciones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Configuracion"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}